Added optional flashcardset, added remove button functionality, edit button pops external view, changed itembox to listbox
Updated constructor with an optional argument, later I'll finish with functionality of storing the data.

Previously:
Made it so you can remove the set by selecting the set and removing it also added edit button which allows you to add flashcard questions and answers (todo: make it so it data is stored somewhere), changed itemboxes to listboxes for easier set tracking (SelectedItem).